reworked timeouts, removed special OSX tests
