--- openwhisk/initHandler_linux_test.go
@@ -1,42 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/*
-This test depends on the fact the proxy can detect the termination
-of an executable that terminates before or after reading the output.
-On OSX command termination is not detected until some input is read.
-*/
-package openwhisk
-
-func Example_badinit_nocompiler() {
-	ts, cur, log := startTestServer("")
-	doRun(ts, "")
-	doInit(ts, "{}")
-	//sys("ls", "_test/exec")
-	doInit(ts, initBinary("_test/exec", ""))      // empty
-	doInit(ts, initBinary("_test/hi", ""))        // say hi
-	doInit(ts, initBinary("_test/hello.src", "")) // source not excutable
-	doRun(ts, "")
-	stopTestServer(ts, cur, log)
-	// Output:
-	// 400 {"error":"no action defined yet"}
-	// 200 {"ok":true}
-	// 400 {"error":"cannot start action: command exited"}
-	// 400 {"error":"cannot start action: command exited"}
-	// 400 {"error":"cannot start action: command exited"}
-	// 400 {"error":"no action defined yet"}
-}
